Refactor: app.js에서 socket server 분리
