+++ BE/ChatApiServer/app.js
@@ -9,7 +9,11 @@ const indexRouter = require('./routes/index');
 const app = require('express')();
 const server = require('http').createServer(app);
 const redis = require('redis');
-const io = require('socket.io')(server);
+
+const store = redis.createClient(6379, 'localhost');
+const pub = redis.createClient(6379, 'localhost');
+const sub = redis.createClient(6379, 'localhost');
+const io = require('./socket/RoomSocket')(server, pub, sub, store);
 
 app.use(cors());
 app.set('view engine', 'ejs');
@@ -35,62 +39,9 @@ if (app.get('env') === 'development') {
 		});
 	});
 }
-
-// 컨텍스트를 공유하기위해, redis store를 바인딩.
-// io.adapter(redis({ host: 'localhost', port: 6379 }));
-let store = redis.createClient(6379, 'localhost');
-let pub = redis.createClient(6379, 'localhost');
-let sub = redis.createClient(6379, 'localhost');
 sub.subscribe('sub');
 sub.on('subscribe', function(channel, count) {
 	console.log('Subscribed to ' + channel + '. Now subsribed to ' + count + 'channel(s).');
-});
-sub.on('message', function(channel, data) {
-	data = JSON.parse(data);
-	if (data.sendType === 'sendToAllClientsInRoom') {
-		io.sockets.to(data.roomId).emit(data.event, { message: data.message, roomId: data.roomId });
-	}
-	if (data.sendType === 'sentToSelf') {
-		io.emit(data.event, data.data);
-	}
-});
-io.on('connection', (socket) => {
-	console.log('a user connected' + socket);
-	socket.on('join', (data) => {
-		let reply = JSON.stringify({
-			event: 'chat',
-			message: 'person JOIN',
-			roomId: data,
-			sendType: 'sendToAllClientsInRoom'
-		});
-		socket.join(data);
-		pub.publish('sub', reply);
-	});
-	socket.on('chat', (data) => {
-		let reply = JSON.stringify({
-			event: 'chat',
-			message: data.message,
-			roomId: data.roomId,
-			sendType: 'sendToAllClientsInRoom'
-		});
-		pub.publish('sub', reply);
-	});
-	socket.on('makeRoom', (data) => {
-		let reply = {
-			roomId: data.roomId,
-			host: data.userName
-		};
-		store.set(reply.roomId, reply.host); // key - value
-		socket.join(reply.roomId);
-	});
-	socket.on('getRoomHostName', (data) => {
-		store.get(data, (err, reply) => {
-			io.emit('getRoomHostName', reply);
-		});
-	});
-	socket.on('disconnect', () => {
-		console.log('disconnected');
-	});
 });
 
 const PORT = 3001;
